refactor(logic): extract remainder faction assignment from game_setup_ownership
Move locked-growth ordering, multi-component DFS packing, and remainder
auto-assignment into a library part file alongside the existing GP packing
part. No behavior change; thins the main ownership module for #2288.

Refs #2288
